chore: major app refactoring
