+++ examples-standalone/grid-charts-integration/src/app/components/footer/footer.component.html
@@ -1,0 +1,6 @@
+<footer class="container-fluid footer text-center d-flex align-items-center">
+    <div class="w-100">
+        <span class="footer-copyright text-center">Copyright © 2019 Progress Software Corporation and/or its subsidiaries or affiliates.</span>
+        <span class="progress-logo d-inline-flex"></span>
+    </div>
+</footer>
