+++ examples-standalone/grid-charts-integration/src/app/components/footer/footer.component.ts
@@ -1,0 +1,11 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+import { Router } from '@angular/router';
+
+@Component({
+    selector: 'app-footer',
+    templateUrl: './footer.component.html',
+    styleUrls: ['./footer.component.scss'],
+    encapsulation: ViewEncapsulation.None
+})
+export class FooterComponent {
+}
