+++ examples-standalone/grid-charts-integration/src/app/components/header/header.component.html
@@ -1,0 +1,11 @@
+<nav class="navbar d-flex align-items-center navbar-expand-lg py-3 header">
+    <div class="container d-flex justify-content-between">
+        <div class="align-self-center">
+            <h1 class="mb-0 header-title">My Stocks Portfolio</h1>
+        </div>
+        <div class="profile-wrapper d-flex flex-column align-self-center">
+            <div class="profile-image"></div>
+            <div class="text-white-50">Collin Johnson</div>
+        </div>
+    </div>
+</nav>
