+++ examples-standalone/grid-charts-integration/src/app/components/header/header.component.ts
@@ -1,0 +1,11 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+
+@Component({
+    selector: 'app-header',
+    templateUrl: './header.component.html',
+    styleUrls: ['./header.component.scss'],
+    encapsulation: ViewEncapsulation.None
+})
+export class HeaderComponent {
+
+}
