+++ examples-standalone/grid-charts-integration/src/app/components/stock-list/stock-list.component.scss
@@ -1,0 +1,69 @@
+@import '../../../styles/variables';
+
+.k-grid {
+  margin-bottom: 100px;
+  border-width: 0;
+
+  td:first-child {
+      text-overflow: unset;
+  }
+
+  .k-grid-header .k-header {
+      vertical-align: middle;
+
+      .grid-header-subtitle {
+          display: block;
+          margin: 0;
+          font-weight: normal;
+          font-size: $grid-header-subtitle;
+      }
+
+      .k-link > .k-icon.k-i-sort-asc-sm,
+      .k-link > .k-icon.k-i-sort-desc-sm {
+          margin-left: $grid-sorting-icon-margin;
+      }
+
+      .grid-header-subtitle + .k-icon.k-i-sort-asc-sm,
+      .grid-header-subtitle + .k-i-sort-desc-sm {
+          position: absolute;
+          top: 50%;
+          margin-top: - #{$grid-sorting-icon-right-position / 2};
+          right: $grid-sorting-icon-right-position;
+      }
+  }
+
+
+  .k-grid-content tr td {
+      vertical-align: top;
+  }
+
+  tr.k-state-selected > td {
+      color: inherit;
+  }
+
+  tr.k-state-selected > td.symbol-col {
+      color: $grid-row-selection-bg;
+  }
+
+  .symbol-col {
+      color: $grid-row-selection-bg;
+  }
+
+  th,
+  .symbol-col,
+  .price-col {
+      font-weight: bold;
+  }
+
+  .grid-cell-positive {
+      color: $grid-cell-positive-color;
+  }
+
+  .grid-cell-negative {
+      color: $grid-cell-negative-color;
+  }
+}
+
+.k-custom-notification {
+  margin: $notification-margin 0;
+}
