--- examples-standalone/grid-charts-integration/src/styles.css
@@ -1,53 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-
-/* .header {
-    height: 140px;
-    background: url(https://telerik.github.io/kendo-angular/finance-portfolio/header-bg.fb71804a4fb1923f2360.svg) bottom no-repeat;
-}
-
-.header h1 {
-    padding-top: .5rem;
-    font-family: Ubuntu,sans-serif;
-    font-weight: 200;
-    color: #fff;
-}
-
-.header .profile-wrapper {
-    outline: 0;
-    cursor: pointer;
-}
-
-.header .profile-image {
-    align-self: center;
-    width: 58px;
-    height: 58px;
-    border-radius: 50%;
-    background-image: url(https://telerik.github.io/kendo-angular/finance-portfolio/user.30f34c6e5c421dc64e98.jpg);
-    background-size: cover;
-    background-position: center;
-}
-
-main {
-  padding: 50px 0;
-  min-height: 100vh;
-  background-color: rgba(236,238,239,.5);
-}
-
-.footer {
-    position: fixed;
-    bottom: 0;
-    padding: 10px;
-    background: url(https://telerik.github.io/kendo-angular/finance-portfolio/footer-bg.a08c1749b859592b8bc4.svg) 0 0/cover no-repeat #00f;
-    z-index: 999;
-    font-size: 16px;
-    color: #fff;
-}
-
-.footer .progress-logo {
-    width: 100px;
-    height: 20px;
-    vertical-align: middle;
-    background-image: url(https://telerik.github.io/kendo-angular/finance-portfolio/progress-logo.86964657182b92a59a85.svg);
-    background-position: center;
-    background-repeat: no-repeat;
-} */
