+++ examples-standalone/grid-charts-integration/src/styles/_kendo.scss
@@ -1,0 +1,25 @@
+// colors
+$accent: #1C7CD5;
+
+// buttons
+$button-bg: white;
+
+// grid
+$grid-padding-x: 0.4rem;
+$grid-padding-y: 0.4rem;
+$grid-alt-bg: transparent;
+$grid-cell-vertical-border-width: 0;
+$grid-cell-horizontal-border-width: 1px;
+$grid-header-bg: #ECEEEF;
+$grid-header-text: #55595C;
+$grid-sorting-index-font-size: 20px;
+
+@import "~@progress/kendo-theme-bootstrap/scss/button";
+@import "~@progress/kendo-theme-bootstrap/scss/grid";
+@import "~@progress/kendo-theme-bootstrap/scss/menu";
+@import "~@progress/kendo-theme-bootstrap/scss/checkbox";
+@import "~@progress/kendo-theme-bootstrap/scss/dropdownlist";
+@import "~@progress/kendo-theme-bootstrap/scss/multiselect";
+@import "~@progress/kendo-theme-bootstrap/scss/dataviz";
+@import "~@progress/kendo-theme-bootstrap/scss/window";
+@import "~@progress/kendo-theme-bootstrap/scss/notification";
