+++ examples-standalone/grid-charts-integration/src/styles/_main.scss
@@ -1,0 +1,10 @@
+@import "./bootstrap";
+@import "./typography";
+@import "./variables";
+@import "./kendo";
+@import "./responsive";
+
+body {
+  font-family: $roboto;
+  background: $body-base-bg;
+}
