+++ examples-standalone/grid-charts-integration/src/styles/_variables.scss
@@ -25,10 +25,8 @@ $font-size-default: 16px;
 
 // header
 $header-height: 140px;
-$header-title-padding: .5rem;
 $profile-image-width: 58px;
 $profile-image-height: 58px;
-$profile-image-margin-top: 2rem;
 
 // footer
 $footer-padding: 10px;
@@ -46,9 +44,6 @@ $dropownlist-add-new-padding: 8px;
 $dropownlist-add-new-width: 125px;
 $dropownlist-add-new-height: 40px;
 
-// splitter
-$splitter-bar-boxshadow: 5px 5px 5px rgba(#000, 0.2);
-
 // grid
 $grid-header-subtitle: 13px;
 $grid-row-selection-bg: #007BFF;
@@ -57,36 +52,12 @@ $grid-cell-negative-color: $negative-value;
 $grid-sorting-icon-margin: $margin-sm;
 $grid-sorting-icon-right-position: 20px;
 
-// buttons
-$nav-button-width: 150px;
-$nav-button-active-bg: $active-state-base;
+// redirect buttons
+$redirect-buttons-padding: 20px;
+$redirect-buttons-right: 25px;
+$redirect-buttons-bottom: 70px;
+$redirect-buttons-border-radius: 5px;
+$redirect-buttons-shadow: 0 2px 6px rgba(0,0,0,0.08);
 
-// badge
-$badge-padding: 10px;
-$badge-height: 22px;
-
-// heatmap
-$heatmap-base-bg: $body-base-bg;
-$heatmap-size-container-mb: -35px;
-$heatmap-scale-wrap-width: 385px;
-$heatmap-scape-block-width: 35px;
-$heatmap-scale-block-height: 18px;
-
-// tooltip
-$tooltip-content-padding-x: 6px;
-$tooltip-content-padding-y: 2 * $tooltip-content-padding-x;
-
-// user profile
-$profile-font-size: 1.8em;
-$profile-current-value-font-size: 1em;
-$profile-table-padding: 8px 12px 8px 0;
-$profile-table-cell-width: 180px;
-$profile-icon-size: 38px;
-$profile-user-wrapper: 210px;
-
-// charts
-$chart-navigation-drag-handle-bg: #ECEEEF;
-$chart-navigation-drag-handle-border: #818A91;
-
-// dateinputs
-$dateinput-width: 140px;
+// notification
+$notification-margin: 30px;
